fix: cast ARITY to usize in PathHash::hash_empty

+++ console/collections/src/kary_merkle_tree/helpers/path_hash.rs
@@ -28,7 +28,7 @@ pub trait PathHash: Clone + Send + Sync {
 
     /// Returns the empty hash.
     fn hash_empty<const ARITY: u8>(&self) -> Result<Self::Hash> {
-        let children = vec![Self::Hash::default(); ARITY];
+        let children = vec![Self::Hash::default(); ARITY as usize];
         self.hash_children(&children)
     }
